DEF-2817 Cleanup of potentially bad error messages (#1714)
* DEF-2817 Added printf format checks on luaL_error
Turned on error when formatting errors are found

* Fixed test code

* DEF-2817 Added dmHashReverseSafe64() for safer error messages

* DEF-2817 Platform formatting fixes

* More platform fixes

* More platform fixes

* More build fixes

+++ engine/rig/src/rig.cpp
@@ -1534,7 +1534,7 @@ namespace dmRig
         }
         uint32_t ik_index = FindIKIndex(instance, constraint_id);
         if (ik_index == ~0u) {
-            dmLogError("Could not find IK constraint (%llu)", constraint_id);
+            dmLogError("Could not find IK constraint (%llu)", (unsigned long long)constraint_id);
             return 0x0;
         }
